Add 'site' field in GeneralSettings
This field is shown only if config key 'show_site_box' is enabled.
Its value is the AD Site where the zentyal server will be placed
when joining a domain.

Also the domain join process has been improved, it seems that since
RC1 there is no need to wait for RID pool allocation.