Fix different uidNumber|gidNumber for objects in the domain controllers
The problem is solver infering the uidNumber and gidNumber from the
samba RID (which is guaranteed to be uniq). This requiere to create the
object in samba database before zentyal LDAP, so new methods are added
to LdapUserBase to notify observers about new objects before they are
created, and the observers can add, modify or delete attributes from the
entry. Also new methods are defined to notify about problems, so observers
can "rollback" their own changes.