Refactor samba antivirus
The zavs VFS plugin had two problems, huge memmory usage and huge
server overload because the scanner was loaded in each connection.

The new architecture has two parts. A VFS plugin (modified version
of scannedonly) sends to a socket (/var/lib/zentyal/zavs) the files
to scan, one in each line, with this format:
    user0x1Eaddr0x1Efile\n
    user - The user who access to the file
    addr - IP address of the client machine
    file - Absolute path of the file to scan
The fields are delimited by the non-printable ASCII char 0x1E.

A new daemon, zavsd creates this socket and listen on it. It has
a pool of threads locked in a queue where the main thread puts the
files to scan.
This new daemon is fully integrated with Zentyal framework and has
its own log file, /var/log/zentyal/zavsd.log