Improve change-hostname, delete all references to current name
This change is required because Zentyal does not manage this file.
At server installation time, a line with a interface IP address can
be inserted in /etc/hosts, causing problems later if IP addresses
are changed in Zentyal interface. This file should only resolve the
host name to addresses assigned to the loopback interface.