Update RemoteSettings.pm
The commit 205a4d0d0694c9513520177db4e3b1cfbbbed488 for #797 has broken my rsync backup configuration which has been working for a year or more on Zentyal 3.0.

Debugging this I tried running the backup manually:
<pre>
/usr/share/zentyal-ebackup/backup-tool --full
Backup module configuration is not completed. Configure it and retry
</pre>

I suggest updating the configurationIsComplete() function to not fail when there is no password defined for the rsync method.