Split openchange config into its own file
This change is required due to different permissions requirements. The
openchange configuration when provisioned using mysql backend require the
password to be written to smb.conf, so it should not be world readable but
ebox group readable. Bind also needs access to smb.conf to read the realm and
build the SOA record, otherwise kerberos will not work correctly as the
primary name server will not contain the domain part. So to avoid permissions
conflicts, the openchange configuration has been split in its own file, and
smb.conf made world readable.