Disable Vacations feature from SOGo webmail
Two reasons for this. One the automatic disable doesn't work
with actual configuration without enabling a cron, follow this
http://www.sogo.nu/bugs/view.php?id=1530

Second one is that this feature if used overrides all the functionality
implemented in OpenChange side for Out Of Office, so better get
rid of it for now till we make them understand each other better.