Changes in management of network cron script to avoid locks
If the frequency of the failover check is too short, or the old
frecuency was, it can overlap the saving changes process and
give various problems.

To avoid this siutation we remove the cron file when we begin to save the
changes and we delay its writng to the last moment.